requirements.txt
requests
python-dotenv
lxml
pyimagedl
yt-dlp
tavily-python
loguru
pydantic>=2.0.0
firecrawl-py
